refactor: member를 order 도메인으로 폴더 이동
