+++ src/main/java/org/prgms/springbootjpa/mission3/domain/MemberRepository.java
@@ -1,7 +1,6 @@
-package org.prgms.springbootjpa.mission3.domain.member;
+package org.prgms.springbootjpa.mission3.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
 }
